ai-local-system/core-ai/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI & ML
transformers==4.37.0
sentence-transformers==2.3.1
torch==2.1.2
accelerate==0.26.1
bitsandbytes==0.42.0
fasttext-wheel==0.9.2

# Vector DB
qdrant-client==1.7.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Utilities
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
python-dotenv==1.0.0
pyyaml==6.0.1
python-multipart==0.0.6

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# HTTP
httpx==0.26.0
aiohttp==3.9.1

# Text Processing
nltk==3.8.1
unidecode==1.3.7
regex==2023.12.25

# Caching
redis==5.0.1
